feat: extend balance_unit across all providers and surface via public API
- Fix Python 2 except syntax (TypeError, ValueError) -> (TypeError, ValueError)
- Change remaining_balance from int to float to preserve precision for EURO
  balances (e.g. 2park EURO accounts store decimals)
- Fix test fixture CURRENCY_CODE value from "MIN" to "MINUTE" (actual API value)
- Add BALANCE_UNIT_TIMES/MINUTE/EURO constants to public models and __init__
- Add balance_units to manifest schema, ProviderManifest and ProviderInfo so
  consumers can query supported units without reading raw manifest files
- Populate balance_unit="MINUTE" in the_hague and dvsportal _map_permit so the
  field is consistent with manifest declarations (not None) for fixed-unit providers
- Add balance_units to all three provider manifests based on actual API behaviour
  (verified against app JS): 2park supports TIMES/MINUTE/EURO per account;
  the_hague and dvsportal always use MINUTE
- Add tests for TIMES and EURO balance types and float precision
- Rename internal _normalize_update_fields -> _normalize_string_list
- Remove duplicate BALANCE_UNIT_* constants from 2park/const.py

+++ src/pycityvisitorparking/provider/2park/api.py
@@ -428,7 +428,9 @@ class Provider(BaseProvider):
         amount = _parse_balance_amount(balance)
         balance_unit = _parse_balance_currency(balance)
         permit_id = self._coerce_id(self._product_id) or "permit"
-        return Permit(id=permit_id, remaining_balance=amount, zone_validity=[], balance_unit=balance_unit)
+        return Permit(
+            id=permit_id, remaining_balance=amount, zone_validity=[], balance_unit=balance_unit
+        )
 
     def _map_reservation_list(self, details: Any) -> list[Reservation]:
         if not isinstance(details, dict):
@@ -577,15 +579,15 @@ class Provider(BaseProvider):
         )
 
 
-def _parse_balance_amount(balance: dict[str, Any]) -> int:
+def _parse_balance_amount(balance: dict[str, Any]) -> float:
     for param in balance.get("ble_parameters", []):
         if param.get("prr_label") == BALANCE_AMOUNT_LABEL:
             raw = param.get("prr_value")
             try:
-                return int(float(raw))
-            except (TypeError, ValueError):
-                return 0
-    return 0
+                return float(raw)
+            except TypeError, ValueError:
+                return 0.0
+    return 0.0
 
 
 def _parse_balance_currency(balance: dict[str, Any]) -> str | None:
